+++ internal/pipe/archiver/archiver.go
@@ -31,7 +31,7 @@ func (p Pipe) Run(ctx *context.Context) error {
 	if err := packager.Compress(ctx.Project.Directory, &buf); err != nil {
 		return err
 	}
-	log.Infof("compressed files to archive of %s", common.ByteCountToHuman(int64(buf.Len())))
+	log.Infof("compressed files to archive of %s", common.ByteCountToHuman(buf.Len()))
 
 	if p.Store {
 		fileName := fmt.Sprintf("%s-%s.tgz", ctx.Project.Metadata.Name, ctx.Project.Metadata.Version)
@@ -45,7 +45,7 @@ func (p Pipe) Run(ctx *context.Context) error {
 		if err != nil {
 			return err
 		}
-		log.Debugf("wrote %s to %s", common.ByteCountToHuman(int64(n)), file.Name())
+		log.Debugf("wrote %s to %s", common.ByteCountToHuman(n), file.Name())
 	}
 
 	ctx.SetValue("archive", &buf)
